# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pumba Api",
        "version": "1.0.2"
    },
    "servers": [
        {
            "url": "http://ovz5.j04713753.0n03n.vps.myjino.ru/public/api/",
            "description": "Pumba"
        }
    ],
    "paths": {
...
        "/updateFcmToken": {
            "post": {
                "tags": [
                    "FCM"
                ],
                "summary": "updateFcmToken",
                "description": "updateFcmToken by api_token, fcm_token",
                "operationId": "updateFcmToken",
                "requestBody": {
                    "description": "fcm",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fcm_token"
                                ],
                                "properties": {
                                    "fcm_token": {
                                        "type": "string",
                                        "example": "OzQ50ke3GElJMNvBZm8uksngp8dqNVYAHqr5CGHN9visYI0TYHg1fFdhsNf8BqTpwqDwXqcPhcxzN3Pj"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Wrong response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "User does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "success": {
                                            "type": "string",
                                            "example": "fcm_token updated success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
        "/sendTestFCMMessage": {
            "post": {
                "tags": [
                    "FCM"
                ],
                "summary": "sendTestFCMMessage",
                "description": "sendTestFCMMessage by body, text",
                "operationId": "sendTestFCMMessage",
                "requestBody": {
                    "description": "fcm",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "body",
                                    "title"
                                ],
                                "properties": {
                                    "body": {
                                        "type": "string",
                                        "example": "Test"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Wrong response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "User does not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "success": {
                                            "type": "string",
                                            "example": "Testing Notification Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Authorization": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}